+++ src/main/kotlin/racingcar/Application.kt
@@ -1,17 +1,17 @@
 package racingcar
 
-import racingcar.domain.usecase.CreateCarUseCase
+import racingcar.domain.usecase.CreateCarsUseCase
 import racingcar.domain.usecase.GetWinnersUseCase
 import racingcar.domain.usecase.PlayRaceUseCase
 import racingcar.ui.*
 
 fun main() {
-    val createCarUseCase = CreateCarUseCase()
+    val createCarsUseCase = CreateCarsUseCase()
     val getWinnersUseCase = GetWinnersUseCase()
     val playRaceUseCase = PlayRaceUseCase()
     val raceInputView = RacingInputView()
     val raceResultView = RacingResultView()
     val controller =
-        RacingCarController(raceInputView, raceResultView, createCarUseCase, getWinnersUseCase, playRaceUseCase)
+        RacingCarController(raceInputView, raceResultView, createCarsUseCase, getWinnersUseCase, playRaceUseCase)
     controller.run()
 }
