+++ src/main/kotlin/racingcar/domain/usecase/CreateCarsUseCase.kt
@@ -2,25 +2,18 @@ package racingcar.domain.usecase
 
 import racingcar.domain.entity.Car
 
-class CreateCarUseCase {
+class CreateCarsUseCase {
     private var anonymityCount = INITIAL_ANONYMITY_COUNT
 
     fun execute(input: String): List<Car> {
         val carNames = input.split(DELIMITER)
-        validateCarNames(carNames)
+        validateUniqueName(carNames)
         return carNames.map { name -> createCar(name) }
     }
 
     private fun createCar(name: String): Car {
-        if (name.trim().isEmpty()) return Car("$ANONYMITY${anonymityCount++}")
+        if (name.isBlank()) return Car("$ANONYMITY${anonymityCount++}")
         return Car(name)
-    }
-
-    private fun validateCarNames(carNames: List<String>) {
-        validateUniqueName(carNames)
-        carNames.forEach { name ->
-            validateCarNamesLength(name)
-        }
     }
 
     private fun validateUniqueName(carNames: List<String>) {
@@ -28,15 +21,10 @@ class CreateCarUseCase {
         require(isDuplicate) { println(DUPLICATE_CARS_ERROR_MESSAGE) }
     }
 
-    private fun validateCarNamesLength(name: String) =
-        require(name.length <= MAX_NAME_COUNT) { println(OVER_NAME_COUNT_MESSAGE) }
-
     companion object {
         private const val INITIAL_ANONYMITY_COUNT = 1
         private const val DELIMITER = ","
         private const val ANONYMITY = "익명"
         private const val DUPLICATE_CARS_ERROR_MESSAGE = "자동차 이름에 중복이 있습니다. 중복을 제거 해주세요"
-        private const val OVER_NAME_COUNT_MESSAGE = "자동차 이름을 5글자 이하로 작성해주세요!"
-        private const val MAX_NAME_COUNT = 5
     }
 }
